requirements.txt
streamlit>=1.32.0
chromadb>=0.4.24
langchain>=0.1.12
langchain-community>=0.0.28
pypdf>=4.1.0
sentence-transformers>=2.5.1
python-dotenv>=1.0.1
beautifulsoup4>=4.12.3
